refactor(evals): move harness into skill as uv scripts
- Convert run_bootstrap.py, run_iteration.py, annotate.py, report.py to self-contained uv scripts with PEP 723 inline deps
- Remove harness/ directory, pyproject.toml, uv.lock (no longer needed)
- Scripts are now executable directly: ./skill/run_bootstrap.py --help
- Each script includes all models inline for portability
- Move ops.configure() into main() to avoid errors on --help
