+++ evals/mirascope_skill/skill/report.py
@@ -1,16 +1,126 @@
+#!/usr/bin/env -S uv run
+# /// script
+# requires-python = ">=3.10"
+# dependencies = [
+#     "pydantic>=2.0",
+#     "pyyaml>=6.0",
+# ]
+# ///
 """Compute and print eval metrics.
 
 Usage:
-    python -m harness.report --bootstrap JSON [--iteration JSON] [--sample YAML]
+    ./report.py --bootstrap RESULTS.json [--iteration RESULTS.json] [--sample SAMPLE.yaml]
 """
 
 from __future__ import annotations
 
 import argparse
-import sys
 from pathlib import Path
 
-from .models import BootstrapResult, EvalSample, IterationResult
+import yaml
+from pydantic import BaseModel, Field
+
+# ---------------------------------------------------------------------------
+# Models
+# ---------------------------------------------------------------------------
+
+
+class QueryExpected(BaseModel):
+    invokes_skill: bool = True
+    output_contains: list[str] = []
+    output_excludes: list[str] = []
+    semantic_requirements: list[str] = []
+
+
+class EvalQuery(BaseModel):
+    id: str
+    text: str
+    specificity: str = ""
+    professionalism: str = ""
+    expected: QueryExpected = Field(default_factory=QueryExpected)
+
+
+class Bootstrap(BaseModel):
+    prompt: str
+    specificity: str = ""
+    professionalism: str = ""
+    expected_capabilities: list[str] = []
+
+
+class SampleMetadata(BaseModel):
+    description: str = ""
+    tags: list[str] = []
+    difficulty: str = "medium"
+
+
+class EvalSample(BaseModel):
+    version: str = "1.0"
+    skill_type: str
+    sample_id: str
+    created_at: str = ""
+    metadata: SampleMetadata = Field(default_factory=SampleMetadata)
+    bootstrap: Bootstrap
+    queries: list[EvalQuery]
+
+    @classmethod
+    def from_yaml(cls, path: str | Path) -> "EvalSample":
+        with open(path) as f:
+            return cls.model_validate(yaml.safe_load(f))
+
+
+class QueryResult(BaseModel):
+    query_id: str
+    orchestrated_input: dict | None = None
+    raw_output: str = ""
+    trace_id: str | None = None
+    span_id: str | None = None
+    error: str | None = None
+
+
+class Annotation(BaseModel):
+    query_id: str
+    acceptable: bool
+    feedback: str = ""
+
+
+class BootstrapResult(BaseModel):
+    sample_id: str
+    program_path: str
+    program_code: str
+    query_results: list[QueryResult]
+    annotations: list[Annotation] = []
+
+    def save(self, path: str | Path) -> None:
+        Path(path).write_text(self.model_dump_json(indent=2))
+
+    @classmethod
+    def load(cls, path: str | Path) -> "BootstrapResult":
+        return cls.model_validate_json(Path(path).read_text())
+
+
+class FoldResult(BaseModel):
+    fold_index: int
+    held_out_query_id: str
+    program_code: str
+    query_result: QueryResult
+    annotation: Annotation | None = None
+
+
+class IterationResult(BaseModel):
+    sample_id: str
+    folds: list[FoldResult]
+
+    def save(self, path: str | Path) -> None:
+        Path(path).write_text(self.model_dump_json(indent=2))
+
+    @classmethod
+    def load(cls, path: str | Path) -> "IterationResult":
+        return cls.model_validate_json(Path(path).read_text())
+
+
+# ---------------------------------------------------------------------------
+# Report functions
+# ---------------------------------------------------------------------------
 
 
 def pct(numerator: int, denominator: int) -> str:
@@ -35,7 +145,6 @@ def print_bootstrap_metrics(bootstrap: BootstrapResult, sample: EvalSample | Non
     print(f"  % Acceptable:     {pct(acceptable, total)}")
     print()
 
-    # Breakdown by specificity x professionalism (requires sample)
     if sample:
         queries_by_id = {q.id: q for q in sample.queries}
         breakdown: dict[tuple[str, str], list[bool]] = {}
@@ -95,7 +204,6 @@ def print_comparison(bootstrap: BootstrapResult, iteration: IterationResult) -> 
     print(f"  Delta:                   {delta:+.1f}pp")
     print()
 
-    # Per-query table
     print("Per-Query Before/After")
     print("-" * 60)
     print(f"  {'Query ID':<12} {'Bootstrap':<12} {'Iteration':<12} {'Change':<10}")
@@ -111,6 +219,11 @@ def print_comparison(bootstrap: BootstrapResult, iteration: IterationResult) -> 
             change = "regressed"
         print(f"  {qid:<12} {b:<12} {i:<12} {change:<10}")
     print()
+
+
+# ---------------------------------------------------------------------------
+# Main
+# ---------------------------------------------------------------------------
 
 
 def main() -> None:
