+++ evals/mirascope_skill/skill/run_bootstrap.py
@@ -1,37 +1,177 @@
+#!/usr/bin/env -S uv run
+# /// script
+# requires-python = ">=3.10"
+# dependencies = [
+#     "mirascope[all]==2.2.2",
+#     "pydantic>=2.0",
+#     "pyyaml>=6.0",
+# ]
+# ///
 """Bootstrap eval: generate a program from a sample, run queries, save results.
 
 Usage:
-    python -m harness.run_bootstrap --sample YAML --output DIR [--model MODEL]
+    ./run_bootstrap.py --sample SAMPLE.yaml --output DIR [--model MODEL]
 """
 
 from __future__ import annotations
 
 import argparse
 import json
+import os
+import subprocess
 import sys
 from pathlib import Path
 
+import yaml
 from mirascope import llm, ops
-from pydantic import BaseModel
+from pydantic import BaseModel, Field
 
-from .models import BootstrapResult, EvalSample, QueryResult
-from .program import get_schema, run_program, validate_program
+# ---------------------------------------------------------------------------
+# Models
+# ---------------------------------------------------------------------------
 
-SKILL_MD = Path(__file__).resolve().parent.parent / "skill" / "SKILL.md"
+
+class QueryExpected(BaseModel):
+    invokes_skill: bool = True
+    output_contains: list[str] = []
+    output_excludes: list[str] = []
+    semantic_requirements: list[str] = []
+
+
+class EvalQuery(BaseModel):
+    id: str
+    text: str
+    specificity: str = ""
+    professionalism: str = ""
+    expected: QueryExpected = Field(default_factory=QueryExpected)
+
+
+class Bootstrap(BaseModel):
+    prompt: str
+    specificity: str = ""
+    professionalism: str = ""
+    expected_capabilities: list[str] = []
+
+
+class SampleMetadata(BaseModel):
+    description: str = ""
+    tags: list[str] = []
+    difficulty: str = "medium"
+
+
+class EvalSample(BaseModel):
+    version: str = "1.0"
+    skill_type: str
+    sample_id: str
+    created_at: str = ""
+    metadata: SampleMetadata = Field(default_factory=SampleMetadata)
+    bootstrap: Bootstrap
+    queries: list[EvalQuery]
+
+    @classmethod
+    def from_yaml(cls, path: str | Path) -> "EvalSample":
+        with open(path) as f:
+            return cls.model_validate(yaml.safe_load(f))
+
+
+class QueryResult(BaseModel):
+    query_id: str
+    orchestrated_input: dict | None = None
+    raw_output: str = ""
+    trace_id: str | None = None
+    span_id: str | None = None
+    error: str | None = None
+
+
+class Annotation(BaseModel):
+    query_id: str
+    acceptable: bool
+    feedback: str = ""
+
+
+class BootstrapResult(BaseModel):
+    sample_id: str
+    program_path: str
+    program_code: str
+    query_results: list[QueryResult]
+    annotations: list[Annotation] = []
+
+    def save(self, path: str | Path) -> None:
+        Path(path).write_text(self.model_dump_json(indent=2))
+
+    @classmethod
+    def load(cls, path: str | Path) -> "BootstrapResult":
+        return cls.model_validate_json(Path(path).read_text())
+
+
+# ---------------------------------------------------------------------------
+# Program helpers
+# ---------------------------------------------------------------------------
+
+
+def _run_uv(program_path: str | Path, *args: str, timeout: int = 120) -> subprocess.CompletedProcess[str]:
+    env = os.environ.copy()
+    return subprocess.run(
+        ["uv", "run", str(program_path), *args],
+        capture_output=True,
+        text=True,
+        timeout=timeout,
+        env=env,
+    )
+
+
+def validate_program(program_path: str | Path) -> tuple[bool, str]:
+    """Run --help and --schema to verify the program is well-formed."""
+    path = Path(program_path)
+    if not path.exists():
+        return False, f"Program file not found: {path}"
+
+    try:
+        result = _run_uv(path, "--help")
+        if result.returncode != 0:
+            return False, f"--help failed: {result.stderr}"
+    except subprocess.TimeoutExpired:
+        return False, "--help timed out"
+
+    try:
+        result = _run_uv(path, "--schema")
+        if result.returncode != 0:
+            return False, f"--schema failed: {result.stderr}"
+        schema = json.loads(result.stdout)
+        if "input" not in schema or "output" not in schema:
+            return False, f"--schema missing 'input' or 'output' keys: {list(schema.keys())}"
+    except subprocess.TimeoutExpired:
+        return False, "--schema timed out"
+    except json.JSONDecodeError as e:
+        return False, f"--schema returned invalid JSON: {e}"
+
+    return True, ""
+
+
+def get_schema(program_path: str | Path) -> dict:
+    """Run --schema and return the parsed JSON."""
+    result = _run_uv(program_path, "--schema")
+    if result.returncode != 0:
+        raise RuntimeError(f"--schema failed: {result.stderr}")
+    return json.loads(result.stdout)
+
+
+def run_program(program_path: str | Path, input_json: str, timeout: int = 120) -> tuple[str, str]:
+    """Run --input and return (stdout, stderr)."""
+    result = _run_uv(program_path, "--input", input_json, timeout=timeout)
+    return result.stdout, result.stderr
+
+
+# ---------------------------------------------------------------------------
+# LLM calls
+# ---------------------------------------------------------------------------
+
+SKILL_MD = Path(__file__).resolve().parent / "SKILL.md"
 DEFAULT_MODEL = "anthropic/claude-sonnet-4-5"
-
-# Initialize Mirascope tracing
-ops.configure()
-ops.instrument_llm()
-
-
-# ---------------------------------------------------------------------------
-# LLM helpers
-# ---------------------------------------------------------------------------
 
 
 class GeneratedProgram(BaseModel):
-    code: str
+    code: str = Field(description="Complete Python source code for the program")
 
 
 @ops.trace(tags=["eval", "bootstrap", "generate"])
@@ -54,12 +194,13 @@ Return the full Python source code. The program must:
 2. Define ProgramInput and ProgramOutput Pydantic models with Field descriptions
 3. Support --help, --schema, and --input CLI flags
 4. Use @ops.trace and @llm.call decorators
-5. Include ops.configure() and ops.instrument_llm() at module level"""
+5. Include ops.configure() and ops.instrument_llm() at module level
+6. Follow all robustness patterns from the skill instructions"""
 
 
 class OrchestrationResult(BaseModel):
-    input_json: dict
-    reasoning: str
+    input_json: dict = Field(description="Structured input matching the program's schema")
+    reasoning: str = Field(description="Explanation of how the query was translated")
 
 
 @ops.trace(tags=["eval", "bootstrap", "orchestrate"])
@@ -93,16 +234,15 @@ def main() -> None:
     parser.add_argument("--model", default=DEFAULT_MODEL, help=f"LLM model (default: {DEFAULT_MODEL})")
     args = parser.parse_args()
 
+    # Initialize tracing (requires MIRASCOPE_API_KEY)
+    ops.configure()
+    ops.instrument_llm()
+
     sample = EvalSample.from_yaml(args.sample)
     output_dir = Path(args.output)
     output_dir.mkdir(parents=True, exist_ok=True)
 
-    # Read skill instructions
     skill_instructions = SKILL_MD.read_text()
-
-    # Override model if specified
-    if args.model != DEFAULT_MODEL:
-        generate_program.__wrapped__.__func__._model = args.model  # type: ignore[attr-defined]
 
     print(f"Generating program for sample: {sample.sample_id}")
 
@@ -111,7 +251,6 @@ def main() -> None:
         response = generate_program(skill_instructions, sample.bootstrap.prompt)
         program_code = response.parse().code
 
-    # Write program
     program_path = output_dir / "program.py"
     program_path.write_text(program_code)
     print(f"Program written to: {program_path}")
@@ -120,7 +259,6 @@ def main() -> None:
     valid, error = validate_program(program_path)
     if not valid:
         print(f"Program validation failed: {error}", file=sys.stderr)
-        # Save partial result even on failure
         result = BootstrapResult(
             sample_id=sample.sample_id,
             program_path=str(program_path),
@@ -148,7 +286,6 @@ def main() -> None:
         print(f"\nRunning query {query.id}: {query.text[:80]}...")
 
         try:
-            # Orchestrate: NL → structured input
             with ops.session(id=f"eval-bootstrap-{sample.sample_id}"):
                 orch_response = orchestrate_query(query.text, input_schema)
                 orch = orch_response.parse()
@@ -162,7 +299,6 @@ def main() -> None:
             continue
 
         try:
-            # Run program
             input_json_str = json.dumps(orchestrated_input)
             stdout, stderr = run_program(program_path, input_json_str)
