+++ evals/mirascope_skill/skill/run_iteration.py
@@ -1,7 +1,16 @@
+#!/usr/bin/env -S uv run
+# /// script
+# requires-python = ">=3.10"
+# dependencies = [
+#     "mirascope[all]==2.2.2",
+#     "pydantic>=2.0",
+#     "pyyaml>=6.0",
+# ]
+# ///
 """Iteration eval: LOO cross-validation with feedback-based program improvement.
 
 Usage:
-    python -m harness.run_iteration --sample YAML --bootstrap JSON --output DIR [--model MODEL]
+    ./run_iteration.py --sample SAMPLE.yaml --bootstrap RESULTS.json --output DIR [--model MODEL]
 
 Requires all bootstrap queries to be annotated.
 """
@@ -10,44 +19,183 @@ from __future__ import annotations
 
 import argparse
 import json
+import os
+import subprocess
 import sys
 from pathlib import Path
 
+import yaml
 from mirascope import llm, ops
-from pydantic import BaseModel
+from pydantic import BaseModel, Field
 
-from .models import (
-    BootstrapResult,
-    EvalSample,
-    FoldResult,
-    IterationResult,
-    QueryResult,
-)
-from .program import get_schema, run_program, validate_program
+# ---------------------------------------------------------------------------
+# Models
+# ---------------------------------------------------------------------------
+
+
+class QueryExpected(BaseModel):
+    invokes_skill: bool = True
+    output_contains: list[str] = []
+    output_excludes: list[str] = []
+    semantic_requirements: list[str] = []
+
+
+class EvalQuery(BaseModel):
+    id: str
+    text: str
+    specificity: str = ""
+    professionalism: str = ""
+    expected: QueryExpected = Field(default_factory=QueryExpected)
+
+
+class Bootstrap(BaseModel):
+    prompt: str
+    specificity: str = ""
+    professionalism: str = ""
+    expected_capabilities: list[str] = []
+
+
+class SampleMetadata(BaseModel):
+    description: str = ""
+    tags: list[str] = []
+    difficulty: str = "medium"
+
+
+class EvalSample(BaseModel):
+    version: str = "1.0"
+    skill_type: str
+    sample_id: str
+    created_at: str = ""
+    metadata: SampleMetadata = Field(default_factory=SampleMetadata)
+    bootstrap: Bootstrap
+    queries: list[EvalQuery]
+
+    @classmethod
+    def from_yaml(cls, path: str | Path) -> "EvalSample":
+        with open(path) as f:
+            return cls.model_validate(yaml.safe_load(f))
+
+
+class QueryResult(BaseModel):
+    query_id: str
+    orchestrated_input: dict | None = None
+    raw_output: str = ""
+    trace_id: str | None = None
+    span_id: str | None = None
+    error: str | None = None
+
+
+class Annotation(BaseModel):
+    query_id: str
+    acceptable: bool
+    feedback: str = ""
+
+
+class BootstrapResult(BaseModel):
+    sample_id: str
+    program_path: str
+    program_code: str
+    query_results: list[QueryResult]
+    annotations: list[Annotation] = []
+
+    def save(self, path: str | Path) -> None:
+        Path(path).write_text(self.model_dump_json(indent=2))
+
+    @classmethod
+    def load(cls, path: str | Path) -> "BootstrapResult":
+        return cls.model_validate_json(Path(path).read_text())
+
+
+class FoldResult(BaseModel):
+    fold_index: int
+    held_out_query_id: str
+    program_code: str
+    query_result: QueryResult
+    annotation: Annotation | None = None
+
+
+class IterationResult(BaseModel):
+    sample_id: str
+    folds: list[FoldResult]
+
+    def save(self, path: str | Path) -> None:
+        Path(path).write_text(self.model_dump_json(indent=2))
+
+    @classmethod
+    def load(cls, path: str | Path) -> "IterationResult":
+        return cls.model_validate_json(Path(path).read_text())
+
+
+# ---------------------------------------------------------------------------
+# Program helpers
+# ---------------------------------------------------------------------------
+
+
+def _run_uv(program_path: str | Path, *args: str, timeout: int = 120) -> subprocess.CompletedProcess[str]:
+    env = os.environ.copy()
+    return subprocess.run(
+        ["uv", "run", str(program_path), *args],
+        capture_output=True,
+        text=True,
+        timeout=timeout,
+        env=env,
+    )
+
+
+def validate_program(program_path: str | Path) -> tuple[bool, str]:
+    """Run --help and --schema to verify the program is well-formed."""
+    path = Path(program_path)
+    if not path.exists():
+        return False, f"Program file not found: {path}"
+
+    try:
+        result = _run_uv(path, "--help")
+        if result.returncode != 0:
+            return False, f"--help failed: {result.stderr}"
+    except subprocess.TimeoutExpired:
+        return False, "--help timed out"
+
+    try:
+        result = _run_uv(path, "--schema")
+        if result.returncode != 0:
+            return False, f"--schema failed: {result.stderr}"
+        schema = json.loads(result.stdout)
+        if "input" not in schema or "output" not in schema:
+            return False, f"--schema missing 'input' or 'output' keys"
+    except subprocess.TimeoutExpired:
+        return False, "--schema timed out"
+    except json.JSONDecodeError as e:
+        return False, f"--schema returned invalid JSON: {e}"
+
+    return True, ""
+
+
+def get_schema(program_path: str | Path) -> dict:
+    result = _run_uv(program_path, "--schema")
+    if result.returncode != 0:
+        raise RuntimeError(f"--schema failed: {result.stderr}")
+    return json.loads(result.stdout)
+
+
+def run_program(program_path: str | Path, input_json: str, timeout: int = 120) -> tuple[str, str]:
+    result = _run_uv(program_path, "--input", input_json, timeout=timeout)
+    return result.stdout, result.stderr
+
+
+# ---------------------------------------------------------------------------
+# LLM calls
+# ---------------------------------------------------------------------------
 
 DEFAULT_MODEL = "anthropic/claude-sonnet-4-5"
 
-# Initialize Mirascope tracing
-ops.configure()
-ops.instrument_llm()
-
-
-# ---------------------------------------------------------------------------
-# LLM helpers
-# ---------------------------------------------------------------------------
-
 
 class ImprovedProgram(BaseModel):
-    code: str
+    code: str = Field(description="Complete improved Python source code")
 
 
 @ops.trace(tags=["eval", "iteration", "improve"])
 @llm.call(DEFAULT_MODEL, format=ImprovedProgram)
-def improve_program(
-    original_code: str,
-    training_examples: str,
-    bootstrap_prompt: str,
-) -> str:
+def improve_program(original_code: str, training_examples: str, bootstrap_prompt: str) -> str:
     return f"""You are improving a Mirascope program based on human feedback.
 
 Here is the original bootstrap request that created this program:
@@ -75,8 +223,8 @@ Return the complete improved Python source code."""
 
 
 class OrchestrationResult(BaseModel):
-    input_json: dict
-    reasoning: str
+    input_json: dict = Field(description="Structured input matching the program's schema")
+    reasoning: str = Field(description="Explanation of how the query was translated")
 
 
 @ops.trace(tags=["eval", "iteration", "orchestrate"])
@@ -97,7 +245,7 @@ Return the structured input as input_json and explain your reasoning."""
 
 
 # ---------------------------------------------------------------------------
-# Main
+# Helpers
 # ---------------------------------------------------------------------------
 
 
@@ -129,6 +277,11 @@ def format_training_examples(bootstrap: BootstrapResult, sample: EvalSample, exc
     return "\n".join(lines)
 
 
+# ---------------------------------------------------------------------------
+# Main
+# ---------------------------------------------------------------------------
+
+
 def main() -> None:
     parser = argparse.ArgumentParser(
         description="Iteration eval: LOO CV with feedback-based improvement"
@@ -138,6 +291,10 @@ def main() -> None:
     parser.add_argument("--output", required=True, help="Output directory for results")
     parser.add_argument("--model", default=DEFAULT_MODEL, help=f"LLM model (default: {DEFAULT_MODEL})")
     args = parser.parse_args()
+
+    # Initialize tracing (requires MIRASCOPE_API_KEY)
+    ops.configure()
+    ops.instrument_llm()
 
     sample = EvalSample.from_yaml(args.sample)
     bootstrap = BootstrapResult.load(args.bootstrap)
@@ -164,10 +321,8 @@ def main() -> None:
         fold_dir = output_dir / "folds" / f"fold_{fold_idx:02d}"
         fold_dir.mkdir(parents=True, exist_ok=True)
 
-        # Build training examples (exclude held-out query)
         training = format_training_examples(bootstrap, sample, exclude_query_id=query.id)
 
-        # Generate improved program
         with ops.session(id=f"eval-{sample.sample_id}-fold-{fold_idx}"):
             response = improve_program(
                 original_code=bootstrap.program_code,
@@ -176,12 +331,10 @@ def main() -> None:
             )
             improved_code = response.parse().code
 
-        # Write improved program
         program_path = fold_dir / "program.py"
         program_path.write_text(improved_code)
         print(f"  Improved program written to: {program_path}")
 
-        # Validate
         valid, error = validate_program(program_path)
         if not valid:
             print(f"  Validation failed: {error}")
@@ -198,7 +351,6 @@ def main() -> None:
 
         print("  Program validated")
 
-        # Get schema and orchestrate held-out query
         try:
             schema = get_schema(program_path)
             input_schema = schema["input"]
@@ -220,7 +372,6 @@ def main() -> None:
             ))
             continue
 
-        # Run held-out query
         try:
             input_json_str = json.dumps(orchestrated_input)
             stdout, stderr = run_program(program_path, input_json_str)
@@ -253,7 +404,6 @@ def main() -> None:
             query_result=qr,
         ))
 
-    # Save results
     result = IterationResult(sample_id=sample.sample_id, folds=folds)
     result_path = output_dir / "iteration_results.json"
     result.save(result_path)
